backend/kosciuszkon/serializers/CalendarSerializer.py
from rest_framework import serializers
from kosciuszkon.models import Calendar

class CalendarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Calendar
        fields = '__all__'

